organized leftbar / made a reusable component for rendering icons & links
