+++ client/src/components/leftBar/leftBarSections.jsx/leftBarEnd/LeftBarEnd.jsx
@@ -1,0 +1,13 @@
+import LeftBarLinks from "../leftBarLinks/LeftBarLinks";
+import { menuItems } from "./menuItems";
+
+const LeftBarEnd = () => {
+  return (
+    <div className="LEFT_BAR__menu">
+      <span>Others</span>
+      <LeftBarLinks menuItems={menuItems} />
+    </div>
+  );
+};
+
+export default LeftBarEnd;
