+++ client/src/components/leftBar/leftBarSections.jsx/leftBarEnd/menuItems.js
@@ -1,0 +1,19 @@
+import { courses, fund, tutorials } from "../../../../assets/img";
+
+export const menuItems = [
+  {
+    icon: fund,
+    label: "Fundraiser",
+    path: "#",
+  },
+  {
+    icon: tutorials,
+    label: "Tutorials",
+    path: "#",
+  },
+  {
+    icon: courses,
+    label: "Courses",
+    path: "#",
+  },
+];
