+++ client/src/components/leftBar/leftBarSections.jsx/leftBarMid/menuItems.js
@@ -1,0 +1,23 @@
+import { events, gaming, gallery, videos } from "../../../../assets/img";
+export const menuItems = [
+  {
+    icon: events,
+    label: "Events",
+    path: "/events",
+  },
+  {
+    icon: gaming,
+    label: "Gaming",
+    path: "/gaming",
+  },
+  {
+    icon: gallery,
+    label: "Gallery",
+    path: "/gallery",
+  },
+  {
+    icon: videos,
+    label: "Videos",
+    path: "/videos",
+  },
+];
