+++ client/src/components/leftBar/leftBarSections.jsx/leftBarTop/menuItems.js
@@ -1,0 +1,29 @@
+import { friends, groups, market, memories, watch } from "../../../../assets/img";
+
+export const menuItems = [
+  {
+    icon: friends,
+    label: "Friends",
+    path: "/friends",
+  },
+  {
+    icon: groups,
+    label: "Groups",
+    path: "/groups",
+  },
+  {
+    icon: market,
+    label: "Marketplace",
+    path: "/marketplace",
+  },
+  {
+    icon: watch,
+    label: "Watch",
+    path: "/watch",
+  },
+  {
+    icon: memories,
+    label: "Memories",
+    path: "/memories",
+  },
+];
